gen-ai-experiments/vector_visualization/requirements.txt
weaviate-client==3.26.7
numpy>=1.24.0
scikit-learn>=1.3.0
bokeh>=3.0.0
plotly